Mol/cc cluster+vae

+++ molecule/preprocess_MoleOOD.py
@@ -8,9 +8,9 @@ import json
 
 from molecule.external_repo.MoleOOD.OGB.modules.ChemistryProcess import get_substructure
 
+CLUSTER_PATH = "/export/livia/datasets/datasets/public/molecule/data"
 
-def get_result_dir(dataset):
-    work_dir = "data/{}".format(dataset)
+def get_result_dir(work_dir):
     result_dir = os.path.join(work_dir, "moleood")
     if not os.path.exists(result_dir):
         os.makedirs(result_dir)
@@ -23,26 +23,7 @@ if __name__ == "__main__":
         "--datasets",
         nargs="+",
         default=[
-            "hERG",
-            "hERG_Karim",
-            "AMES",
-            "DILI",
-            "Carcinogens_Lagunin",
-            "Tox21",
-            "ClinTox",
-            "PAMPA_NCATS",
-            "HIA_Hou",
-            "Pgp_Broccatelli",
-            "Bioavailability_Ma",
-            "BBB_Martins",
-            "CYP2C19_Veith",
-            "CYP2D6_Veith",
-            "CYP3A4_Veith",
-            "CYP1A2_Veith",
-            "CYP2C9_Veith",
-            "CYP2C9_Substrate_CarbonMangels",
-            "CYP2D6_Substrate_CarbonMangels",
-            "CYP3A4_Substrate_CarbonMangels",
+            "BindingDB_Ki",
         ],
         help="the datasets to preprocess",
     )
@@ -58,24 +39,36 @@ if __name__ == "__main__":
         default="brics",
         help="the method to decompose the molecule, brics or recap",
     )
+    parser.add_argument(
+        "--data-path",
+        type=str,
+        default=CLUSTER_PATH if os.path.exists(CLUSTER_PATH) else "data",
+        help="Path to the data folder",
+    )
+
     args = parser.parse_args()
     print(args)
     for dataset in args.datasets:
-        result_dir = get_result_dir(dataset)
-        if not os.path.exists(result_dir):
-            os.mkdir(result_dir)
+        data_path = os.path.join(args.data_path, dataset)
+        if os.path.exists(f"{data_path}/moleood/substructures.pkl"):
+            print(f"Dataset {dataset} already preprocessed, skipping")
+            continue
+        if os.path.exists(f"{data_path}/smiles.json"):
+            result_dir = get_result_dir(data_path)
+            if not os.path.exists(result_dir):
+                os.mkdir(result_dir)
 
-        with open(f"data/{dataset}/smiles.json", "r") as f:
-            smiles = json.load(f)
+            with open(f"{data_path}/smiles.json", "r") as f:
+                smiles = json.load(f)
 
-        file_name = (
-            "substructures.pkl" if args.method == "brics" else "substructures_recap.pkl"
-        )
-        file_name = os.path.join(result_dir, file_name)
-        substruct_list = []
-        for idx, smile in enumerate(tqdm(smiles)):
-            tx = get_substructure(smile=smile, decomp=args.method)
-            substruct_list.append(tx)
+            file_name = (
+                "substructures.pkl" if args.method == "brics" else "substructures_recap.pkl"
+            )
+            file_name = os.path.join(result_dir, file_name)
+            substruct_list = []
+            for idx, smile in enumerate(tqdm(smiles)):
+                tx = get_substructure(smile=smile, decomp=args.method)
+                substruct_list.append(tx)
 
-        with open(file_name, "wb") as Fout:
-            pickle.dump(substruct_list, Fout)
+            with open(file_name, "wb") as Fout:
+                pickle.dump(substruct_list, Fout)
